Rewrote & optimized a few tools
Prior to this commit, the move and paint tools were structured/written
very differently than the rest of the tools, given that they were the
first ones created for this project. They now follow a similar structure
to the new ones.

Also, the move and rotate tools (which both used bounding boxes for
their global and center modes) were causing a lot of lag when handling
large objects; this is no longer the case.