Selection operations are ignored when dragging
Prior to this commit, if you were using a tool that had handles and
accidentally missed the handles, when you let go of the mouse, it would
mess up your selection and not do what you want. To deal with that
issue, selection operations (including selecting and deselecting) are
now ignored when the mouse position at which the click was started
differs from that at which it is released.