Fix the issue where the 2D selection system gets stuck
Instead of keeping a variable that refers to the camera
that was on when the tool was created, this patch makes
anything that needs the camera access it directly. This
prevents issues where the camera was replaced with
another and the platform begins to freak out about the
camera not existing, or selecting the wrong items in
2D selection mode (because it uses the orientation info
from the wrong, outdated camera).