Fix the bug that frequently prevents the tool from properly equipping
Given that there were so many UI components, it took a lengthy amount
of time to replicate all of them to the client. The tool expected them
to already be there when making a copy of the base GUI.

This patch solves the issue by providing a total count of the number of
UI components that should replicate to the client, and having the client
wait until the number of expected UI components is fulfilled--it allows it
to know that all the required UI components have arrived and that it can
proceed safely without missing any.