Update Sizes and FormFactors on NewPart tool
I changed the size of newly created spawns and wedges from the default 4, 1.2, 2 to 4, 1, 2 (see lines 86 and 101). I also added "NewPart.FormFactor = "Custom";" to all the types (see line 104). These changes will allow for more seamless building, and will allow all of the parts to be on the same playing field, per se.