Fix the way the plugins check if Studio is in solo mode
Previously, they checked for the existence of a Visit instance
under Game, but that instance is also added when opening the place
from green Edit buttons, such that it's not a reliable way of
knowing when it's in Solo or not. There's no official way to
check this, so waiting a while for players may be the best,
most accurate way to go for the moment.

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>