Add Custom Part Templates
Implementation of #13

So essentially, if a user right-clicks a part with this tool, it'll create a new option on the dropdown which will be the template created. It isn't very good at the moment (e.g. doesn't allow naming or deleting of templates), but it's on it's way, and it's the best I can do with my current knowledge of the guis. I don't know how to make one, but a dialog box like "Would you like to make this part into a template?" would be nice.