Add hotkeys for solid modeling capability
Gigs himself guaranteed an auto-approval for this lol. Supplies hotkeys (without buttons because idk how the GUI works yet) for unioning, separating, and negating, and I got the history into the union but it's 1:20am so I didn't try to figure it out yet for separate and I'm too lazy to do it right now for negate.