Duplicate detection
Implementation of #29. When a part is created, its Archivable property is set to false. If it isn't changed, it doesn't save. Simple enough. This'll work great so long as the export doesn't export parts without Archivable set to true, but I can't find the code for it to change it.