Restrict operations dependent on initial state to known parts
This change makes the Move, Resize, & Rotate tools apply change
operations that are dependent on initial state only to parts
for which the initial state is known. It should fix cases where
parts enter the selection in the middle of these operations and
cause the tool to fail due to attempting to act upon the new
parts without knowing their initial state.

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>