Fix automatic updating by using a module to load the new tool
This change switches the loading method for autoupdating the tool
to requiring a module containing the latest tool asset. It should
resolve issues with InsertService not being able to load the tool
in group places due to not owning the asset, since modules can be
required from any place regardless of whether they are owned by
the place owner or not.

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>