Fix tool replacement in automatic updating
This change fixes an issue where the tool was not able to properly
disable due to being destroyed too early, causing connections from
previous versions to stick around and cause problems while using
the updated tool. It also fixes a problem in the loading sequence
where the tool hangs upon not being able to load due to waiting on
the autoupdating script, which is counted as a descendant but
removes itself when updating the tool.

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>