Prevent tool enabling and disabling sequences from overlapping
This patch fixes an issue where the tool can disable midway into
the process of enabling, causing an incomplete clean-up of state
which leads to issues such as hotkeys triggering twice, or parts
of the tool appearing even though it is supposed to be disabled.
The tool now properly responds based on the current state, which
can vary between enabled, enabling, disabling, or disabled, and
may hold requests to enable or disable until the current state's
sequence has finished (e.g. enabling -> enabled, disabling ->
disabled).

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>