Fix dragging issue in Move Tool causing parts to stick to mouse
This patch ensures that the same part targeted immediately upon
clicking is used for the rest of the dragging process (in case of
any delays, which may cause a different part to be assumed as the
target if the mouse position changes during a delay). It also
ensures that the lifecycle of dragging operations is not
interrupted, such that proper cleanup occurs each time.

Signed-off-by: Robert Chiquini <gigsd4x@gmail.com>