Fixed CreatePart and Clone failing to work under Content Streaming By using tags, we can wait for the parts to replicate in. Right now, F3X falters at creating/cloning parts with content streaming. This is because the code assumes that the instances will instantly replicate, which is no longer true with content streaming. Hence, the list of parts the client receives from the server becomes empty, despite being created on the server.
So, from the server, we give the client a tag to listen to for the newly
made parts. Most times, it shouldn't need to wait for the parts to come.
But, with enough parts being cloned, it may need to wait for several
chunks. However, sometimes the parts will never stream in if far
enough, so simply best to give up after a reasonable amount of time.